feat: reset chain config code
