+++ src/pages/components/index.js
@@ -1,2 +1,1 @@
-export {default as ChainIcon} from './ChainIcon'
 export {default as TokenNameAndIcon} from './TokenNameAndIcon'
